# docs/swagger.yaml
basePath: /api
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a learning management system API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Labpro API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with identifier (username/email) and password
      parameters:
      - description: Login credentials
        in: body
        name: login
        required: true
        schema:
          properties:
            identifier:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  token:
                    type: string
                  username:
                    type: string
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
      summary: User login
      tags:
      - auth
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Registration data
        in: body
        name: register
        required: true
        schema:
          properties:
            confirm_password:
              type: string
            email:
              type: string
            first_name:
              type: string
            last_name:
              type: string
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                properties:
                  token:
                    type: string
                  username:
                    type: string
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
      summary: User registration
      tags:
      - auth
  /auth/self:
    get:
      description: Get the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  email:
                    type: string
                  first_name:
                    type: string
                  id:
                    type: string
                  last_name:
                    type: string
                  role:
                    type: string
                  username:
                    type: string
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"